requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=1.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
pymysql>=1.1.0
pydantic>=2.0.0
faiss-cpu>=1.7.4
